feat: 찜 목록조회 추가(미완),@Table(name = "likes") 추가, jpa 타입 Long으로 수정

+++ src/main/java/com/poku/graypants/domain/like/persistence/LikeRepository.java
@@ -8,7 +8,6 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface LikeRepository extends JpaRepository<Like, Integer> {
-
+public interface LikeRepository extends JpaRepository<Like, Long> {
 
 }
